refactor: extract Key Vault secret cache to lib/keyvaultCache.ts
Move refreshKeyVaultSecrets, the per-environment secret cache, the
stale-fetch sequence guard and the env-scoped invalidation out of
App.svelte into src/lib/keyvaultCache.ts. App.svelte keeps only the
activeEnvironment subscription and passes the module's reset/refresh
hooks to openFolderByPath. No behavior change.
